requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
plotly>=5.0.0
huggingface_hub>=0.20.0
python-dotenv>=1.0.0
